Combine password logic into a single controller and add copy tweaks / email layout
