+++ controllers/user/views/emails/forgotten-password.njk
@@ -1,0 +1,11 @@
+{% extends "views/layouts/emails.njk" %}
+
+{% block content %}
+    <h1>The National Lottery Community Fund</h1>
+    <h2>Password reset requested</h2>
+    <p>We recently received a request to reset the password associated with this account. If you didn't initiate this request, please ignore this email.</p>
+    <p>If you did wish to reset the password associated with <strong>{{ email }}</strong>, then please click the following link where you'll be promoted to enter a new password:</p>
+    <p><a href="{{ resetUrl }}">{{ resetUrl }}</a></p>
+    <p>Thank you,</p>
+    <p>The National Lottery Community Fund</p>
+{% endblock %}
