+++ controllers/user/views/forgotten-password.njk
@@ -15,29 +15,32 @@
                 </div>
             {% endif %}
 
-            <form action="" method="post" autocomplete="off">
-                {{ formErrors(errors = errors, title = "There was a problem with your submission") }}
+            {% if not passwordWasJustReset %}
 
-                {% if csrfToken %}
-                    <input type="hidden" name="_csrf" value="{{ csrfToken }}">
-                {% endif %}
-                
+                <form action="" method="post" autocomplete="off">
+                    {{ formErrors(errors = errors, title = "There was a problem with your submission") }}
 
-                {{ formField({
-                    type: 'email',
-                    name: 'username',
-                    label: 'Email address',
-                    explanation: 'Enter the email address associated with your account.',
-                    isRequired: true,
-                    value: formValues['username']
-                }, errors = errors) }}
+                    {% if csrfToken %}
+                        <input type="hidden" name="_csrf" value="{{ csrfToken }}">
+                    {% endif %}
 
-                <div class="form-actions">
-                    <input type="submit" class="btn" value="Reset password" />
-                </div>
 
-                <p><small>Already reset your password? <a href="{{ localify('/user/login') }}">Log in</a></small></p>
-            </form>
+                    {{ formField({
+                        type: 'email',
+                        name: 'username',
+                        label: 'Email address',
+                        explanation: 'Enter the email address associated with your account.',
+                        isRequired: true,
+                        value: formValues['username']
+                    }, errors = errors) }}
+
+                    <div class="form-actions">
+                        <input type="submit" class="btn" value="Reset password" />
+                    </div>
+
+                    <p><small>Already reset your password? <a href="{{ localify('/user/login') }}">Log in</a></small></p>
+                </form>
+            {% endif %}
         </div>
     </main>
 {% endblock %}
