+++ controllers/user/views/login.njk
@@ -8,6 +8,7 @@
     <main role="main" id="content">
         <div class="content-box u-inner-wide-only">
             <h1>{{ title }}</h1>
+            <p>If you already have an account for The National Lottery Community Fund's website, you can log into it here.</p>
 
             {% if alertMessage %}
                 <div class="message" role="alert">
@@ -42,7 +43,7 @@
                     or <a href="/user/register">register</a>
                 </div>
 
-                <p><small><a href="/user/forgotten-password">Forgotten your password?</a></small></p>
+                <p><small><a href="/user/password/forgot">Forgotten your password?</a></small></p>
             </form>
         </div>
     </main>
