+++ controllers/user/views/register.njk
@@ -8,6 +8,7 @@
     <main role="main" id="content">
         <div class="content-box u-inner-wide-only">
             <h1>{{ title }}</h1>
+            <p>You can create an account for use on The National Lottery Community Fund's website here.</p>
 
             {% if alertMessage %}
                 <div class="message" role="alert">
