+++ controllers/user/views/reset-password-expired.njk
@@ -8,7 +8,7 @@
         <div class="content-box u-inner-wide-only">
             <h1>{{ title }}</h1>
             <div class="message" role="alert">
-                Your password reset link was invalid or has expired. <a href="/user/forgotten-password">Please try again</a>.
+                Your password reset link was invalid or has expired. <a href="/user/password/forgot">Please try again</a>.
             </div>
         </div>
     </main>
