+++ controllers/user/views/reset-password.njk
@@ -8,6 +8,7 @@
     <main role="main" id="content">
         <div class="content-box u-inner-wide-only">
             <h1>{{ title }}</h1>
+            <p>Please choose a new password for your account by entering it below. Ensure you choose a strong password.</p>
 
             {% if alertMessage %}
                 <div class="message" role="alert">
